backend-service/ai-model-tester/requirements.txt
flask==2.3.3
firebase-admin==6.2.0
google-generativeai==0.3.1
python-dotenv==1.0.0